refactor(workflow): move runtime host into gateway
