+++ src-tauri/src/adaptor/gateway/workflow/diagnostics.rs
@@ -3,11 +3,11 @@ use crate::adaptor::gateway::workflow::domain_mapping::workflow_definition_to_do
 use crate::adaptor::gateway::workflow::facet::{self, FacetKind};
 use crate::adaptor::gateway::workflow::schema::{NodeDefinition, Rule, WorkflowDefinitionYaml};
 use crate::adaptor::gateway::workflow::span_map::{DiagnosticSpan, YamlSpanMap};
+use crate::adaptor::gateway::workflow::workflow_host::prompt_rendering;
 use crate::domain::workflow::validation;
 use crate::domain::workflow::validation::{
     InvalidArtifactReferenceKind, InvalidRuleKind, InvalidSchemaKind,
 };
-use crate::infrastructure::runtime::workflow_host::prompt_rendering;
 use serde::Serialize;
 use std::collections::{BTreeSet, HashMap, HashSet};
 use std::path::Path;
