+++ src-tauri/src/adaptor/gateway/workflow/workflow_host/lifecycle_commands.rs
