--- src-tauri/src/infrastructure/runtime/mod.rs
@@ -1,3 +1,0 @@
-//! Concrete application runtime integrations.
-
-pub(crate) mod workflow_host;
